Refactor and enhance Bezier curve demo
The curves are drawn based on the number of points now,
so linear, quad, and cubic curves are treated separately.

Curve offset handling is moved out to a new parent scene which
controls the individual offsets of each curve. The offset is updated
automatically, but can also switch to manual mode now.